workaround

+++ gensim/models/keyedvectors.py
@@ -2019,7 +2019,7 @@ def prep_vectors(target_shape, prior_vectors=None, seed=0, dtype=REAL):
         return prior_vectors
     target_count, vector_size = target_shape
     rng = np.random.RandomState(seed)
-    new_vectors = rng.random(target_shape, dtype=dtype)  # [0.0, 1.0)
+    new_vectors = rng.random(target_shape).astype(dtype)  # [0.0, 1.0)
     new_vectors *= 2.0  # [0.0, 2.0)
     new_vectors -= 1.0  # [-1.0, 1.0)
     new_vectors /= vector_size
